Removed the code from init.py files for the translation into the utils and constant files
